requirements.txt
numpy==1.22.4
scipy
tqdm
lpips
mmcv==1.7.0
imageio
imageio-ffmpeg
opencv-python
open3d
trimesh
transformers==4.30.2
diffusers==0.12.1
omegaconf
tensorboard
modelcards
accelerate
einops-exts
entmax
ftfy
ninja
triton
xatlas
smplx
chumpy==0.70
pytorch_lightning==2.0.1
open_clip_torch

